Convert console project to class library and add WinForms UI
