+++ DCCollections.Gui/Form1.cs
@@ -1,10 +1,52 @@
+using Microsoft.Extensions.Configuration;
+
 namespace DCCollections.Gui
 {
     public partial class Form1 : Form
     {
+        private readonly RMCollectionProcessor.CollectionService _service;
+        private readonly Microsoft.Extensions.Configuration.IConfiguration _config;
+
         public Form1()
         {
             InitializeComponent();
+
+            _config = new Microsoft.Extensions.Configuration.ConfigurationBuilder()
+                .SetBasePath(AppContext.BaseDirectory)
+                .AddJsonFile("appsettings.json", optional: true)
+                .Build();
+            _service = new RMCollectionProcessor.CollectionService();
+        }
+
+        private void btnParse_Click(object sender, EventArgs e)
+        {
+            using var ofd = new OpenFileDialog();
+            if (ofd.ShowDialog() == DialogResult.OK)
+            {
+                try
+                {
+                    var result = _service.ParseFile(ofd.FileName);
+                    MessageBox.Show($"Parsed {result.Length} records.", "Success");
+                }
+                catch (Exception ex)
+                {
+                    MessageBox.Show(ex.Message, "Error");
+                }
+            }
+        }
+
+        private void btnGenerate_Click(object sender, EventArgs e)
+        {
+            try
+            {
+                int day = (int)nudDay.Value;
+                var file = _service.GenerateFile(day, _config);
+                MessageBox.Show($"File generated: {file}", "Success");
+            }
+            catch (Exception ex)
+            {
+                MessageBox.Show(ex.Message, "Error");
+            }
         }
     }
 }
